+++ test/typescript/keyboard.test.ts
@@ -19,6 +19,32 @@
 import { test, expect } from "@jest/globals";
 import { Keyboard } from "../../dist/lib/firebolt-manage";
 
+class KBProvider implements Keyboard.KeyboardInputProvider {
+  standard(
+    parameters: Keyboard.KeyboardParameters,
+    session: Keyboard.FocusableProviderSession
+  ): Promise<Keyboard.KeyboardResult> {
+    return Promise.resolve(null);
+  }
+  password(
+    parameters: Keyboard.KeyboardParameters,
+    session: Keyboard.FocusableProviderSession
+  ): Promise<Keyboard.KeyboardResult> {
+    return Promise.resolve(null);
+  }
+  email(
+    parameters: Keyboard.KeyboardParameters,
+    session: Keyboard.FocusableProviderSession
+  ): Promise<Keyboard.KeyboardResult> {
+    return Promise.resolve(null);
+  }
+}
+
+test("Keyboard.provide() declarations", () => {
+  Keyboard.provide("xrn:firebolt:capability:input:keyboard", new KBProvider());
+  expect(1).toBe(1);
+});
+
 test("Keyboard.provide() with blank object", () => {
   expect(() => {
     Keyboard.provide("xrn:firebolt:capability:input:keyboard", {});
