+++ test/typescript/pinChallenge.test.ts
@@ -19,6 +19,23 @@
 import { test, expect } from "@jest/globals";
 import { PinChallenge } from "../../dist/lib/firebolt-manage";
 
+class PCProvider implements PinChallenge.ChallengeProvider {
+  challenge(
+    parameters: object,
+    session: PinChallenge.FocusableProviderSession
+  ): Promise<PinChallenge.PinChallengeResult> {
+    return Promise.resolve(null);
+  }
+}
+
+test("PinChallenge.provide() declarations", () => {
+  PinChallenge.provide(
+    "xrn:firebolt:capability:usergrant:pinchallenge",
+    new PCProvider()
+  );
+  expect(1).toBe(1);
+});
+
 test("PinChallenge.provide() with blank object", () => {
   expect(() => {
     PinChallenge.provide("xrn:firebolt:capability:usergrant:pinchallenge", {});
