+++ test/typescript/privacy.test.ts
@@ -1,23 +1,51 @@
-/*
- * Copyright 2021 Comcast Cable Communications Management, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
 import { test, expect } from "@jest/globals";
 import { Privacy } from "../../dist/lib/firebolt-manage";
+
+class ContentPrivacyProvider implements Privacy.ContentPolicyProvider {
+  contentPolicy(
+    parameters: object,
+    session: Privacy.ProviderSession
+  ): Promise<Privacy.ContentPolicy> {
+    return Promise.resolve(null);
+  }
+}
+
+class AdvertisingPrivacyProvider implements Privacy.AdvertisingPolicyProvider {
+  advertisingPolicy(
+    parameters: object,
+    session: Privacy.ProviderSession
+  ): Promise<object> {
+    return Promise.resolve(null);
+  }
+}
+
+test("Content Policy Provider", () => {
+  Privacy.provide(
+    "xrn:firebolt:capability:privacy:content",
+    new ContentPrivacyProvider()
+  );
+  expect(1).toBe(1);
+});
+
+test("Advertising Policy Provider", () => {
+  Privacy.provide(
+    "xrn:firebolt:capability:privacy:advertising",
+    new AdvertisingPrivacyProvider()
+  );
+  expect(1).toBe(1);
+});
+
+let listenerId: number;
+
+test("Able to get TypeScript listenerId", () => {
+  return Privacy.listen("limitAdTrackingChanged", () => {}).then(
+    (id: number) => {
+      listenerId = id;
+      console.log(listenerId);
+      expect(listenerId > 0).toBe(true);
+    }
+  );
+});
 
 test("privacy.shareWatchHistory()", () => {
   return Privacy.shareWatchHistory(String).then((res: number) => {
@@ -35,15 +63,6 @@ test("privacy.rememberWatchedPrograms()", () => {
   return Privacy.rememberWatchedPrograms(String).then((res: number) => {
     expect(res > 0).toBe(true);
   });
-});
-
-test("privacy.listen()", () => {
-  Privacy.once;
-  return Privacy.listen("requestContentPolicy", () => {}).then(
-    (res: number) => {
-      expect(res > 0).toBe(true);
-    }
-  );
 });
 
 test("privacy.once()", () => {
